wayland: Optimize wayland code
 - Minimize wl_shm_pool profligacy by doing pixel math on a shared memory pool
   instead of creating one for each surface (i.e use a single backing buffer).
 - Switch to 'way_' instead of 'wl_' as an internal prefix to keep
   names distinct from wayland library objects

This significantly reduces memory usage and should improve performance, though
there is still room for improvement (batching/caching draw ops).

+++ src/grid.c
@@ -17,8 +17,8 @@ static void draw_grid(screen_t scr,
 {
 	int i;
 
-	const double ygap = (h - ((nr+1)*sz))/(double)nr;
-	const double xgap = (w - ((nc+1)*sz))/(double)nc;
+	const int ygap = (h - ((nr+1)*sz))/nr;
+	const int xgap = (w - ((nc+1)*sz))/nc;
 
 	if (xgap < 0 || ygap < 0)
 		return;
